caracteristicas: -- Solución de bugs en modulo de juegos -- se agrego el contador de llamitas en las preguntas de la aventura

+++ lib/graphql-config/function_graphql/mutations.dart
@@ -508,9 +508,10 @@ Future logout() async {
   final GoogleSignIn googleSignIn = GoogleSignIn();
   await googleSignIn.signOut();
   final deviceInfo = await PreferencesManager().getDeviceInfo();
+  final token = await PreferencesManager().getUserToken();
   String userId = await PreferencesManager().getUserId();
   // eliminamos el dispositivo del usuario
-  if (GraphQLConfig.development) {
+  if (GraphQLConfig.development && userId.isNotEmpty ) {
     final responseDeleteDevice =
         await deleteDevice(userId, deviceInfo?['deviceId'] ?? '');
